% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{aborting_ctrls}
\alias{aborting_ctrls}
\title{Prepare the aborting controls. The aborting controls are used to stop the training of unpromising models.
If the MSE is still above \code{mse_value} value at iteration \code{check_iter} then the training is stopped.}
\usage{
aborting_ctrls(mse_value = Inf, check_iter = Inf)
}
\arguments{
\item{mse_value}{Value to compare the MSE to. Default: Inf.}

\item{check_iter}{Iteration at which the check is done. Default: Inf.}
}
\value{
aborting_controls
}
\description{
Prepare the aborting controls. The aborting controls are used to stop the training of unpromising models.
If the MSE is still above \code{mse_value} value at iteration \code{check_iter} then the training is stopped.
}
