% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_roxygen.R
\docType{data}
\name{data_mixedml}
\alias{data_mixedml}
\title{The dataset used to play with the package.}
\usage{
data_mixedml
}
\description{
The dataset used to play with the package.
}
\keyword{datasets}
