% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{.get_model_snapshot}
\alias{.get_model_snapshot}
\title{This function captures the current model state in the training loop
the names of the saved variables is defined in MIXEDML_COMPONENTS}
\usage{
.get_model_snapshot()
}
\description{
This function captures the current model state in the training loop
the names of the saved variables is defined in MIXEDML_COMPONENTS
}
