% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlme_tools.R
\name{.initiate_random_hlme}
\alias{.initiate_random_hlme}
\title{Function to plot something}
\usage{
.initiate_random_hlme(random_spec, data, subject, var.time, hlme_controls)
}
\arguments{
\item{random_spec}{random_spec}

\item{data}{data}

\item{subject}{subject}

\item{var.time}{var.time}

\item{hlme_controls}{hlme_controls}
}
\value{
description
}
\description{
Function to plot something
}
