% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlme_tools.R
\name{.predict_with_past_info}
\alias{.predict_with_past_info}
\title{This function computes the predictions of a HLME model using only the past information to compute the random effects
and correlation components. The actual X values are of course still used to compute the estimates.}
\usage{
.predict_with_past_info(hlme_model, data, nproc = 1)
}
\arguments{
\item{hlme_model}{HLME model from the LCMM package}

\item{data}{Data to be used for the prediction. It must have the same format as the one used to fit the hlme model.}

\item{nproc}{Number of processes to use. Default: 1}
}
\description{
This function computes the predictions of a HLME model using only the past information to compute the random effects
and correlation components. The actual X values are of course still used to compute the estimates.
}
