% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{earlystopping_ctrls}
\alias{earlystopping_ctrls}
\title{Prepare the earlystopping controls. The earlystopping controls are used to stop the training
when convergence is reached.
If no improvement of the MSE greater than \code{min_mse_gain} is observed during \code{patience} iterations,
then the training is stopped.
If a validation dataset is provided, the MSE on the validation dataset is used for the early stopping.
Otherwise, the MSE on the training dataset is used.
By default, early stopping is disabled.}
\usage{
earlystopping_ctrls(patience = Inf, min_mse_gain = 0)
}
\arguments{
\item{patience}{Number of iterations without improvement before the training is stopped. Default: Inf.}

\item{min_mse_gain}{Minimal difference of MSE to consider an improvement.
\code{min_mse_gain=1.} means an improvement of at least 1. of the MSE is necessary. Default: 0.}
}
\value{
earlystopping_controls
}
\description{
Prepare the earlystopping controls. The earlystopping controls are used to stop the training
when convergence is reached.
If no improvement of the MSE greater than \code{min_mse_gain} is observed during \code{patience} iterations,
then the training is stopped.
If a validation dataset is provided, the MSE on the validation dataset is used for the early stopping.
Otherwise, the MSE on the training dataset is used.
By default, early stopping is disabled.
}
