% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_model.reservoir.R
\name{ensemble_ctrls}
\alias{ensemble_ctrls}
\title{Prepare the ensemble_controls}
\usage{
ensemble_ctrls(
  seed_list = c(1, 2, 3),
  aggregator = "median",
  scaler = "standard",
  n_procs = 1,
  return_individual = FALSE
)
}
\arguments{
\item{seed_list}{List of seeds used to generate the Reservoir. Default:  c(1, 2, 3)}

\item{aggregator}{Function used to aggregate the predictions of each ESN.
"mean" or "median". Default: "median"}

\item{scaler}{scikit-learn scaler to use on the X data.
"standard", "robust", "min-max", "max-abs". Default: "standard"}

\item{n_procs}{Number of processor to use. 1 means no multiprocessing. Default: 1.}

\item{return_individual}{return predictions for each reservoir (seed). Usefull for random research. Default: FALSE}
}
\value{
ensemble_controls
}
\description{
Prepare the ensemble_controls
}
