% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{fit_ctrls}
\alias{fit_ctrls}
\title{Prepare the fit_controls}
\usage{
fit_ctrls(warmup = 0)
}
\arguments{
\item{warmup}{Number of timesteps to consider as warmup and discard at the beginning. Defalut: 0
of each timeseries before training.}
}
\value{
fit_controls
}
\description{
Please see the
\href{https://reservoirpy.readthedocs.io/en/latest/api/generated/reservoirpy.nodes.ESN.html#reservoirpy.nodes.ESN.fit}{documentation}
of ReservoirPy
}
