% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{get_loglik}
\alias{get_loglik}
\title{Return the negative log-likelihood of the model regarding new data}
\usage{
get_loglik(model, data)
}
\arguments{
\item{model}{Trained MixedML model}

\item{data}{New data (same format as the one used for training)}
}
\value{
Negative log-likelihood of the HLME model
}
\description{
Return the negative log-likelihood of the model regarding new data
}
