% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlme_tools.R
\name{hlme_ctrls}
\alias{hlme_ctrls}
\title{Prepare the hlme_controls}
\usage{
hlme_ctrls(
  cor = NULL,
  idiag = FALSE,
  maxiter = 500,
  nproc = 1,
  B_rand = NULL,
  convB = 1e-04,
  convL = 1e-04,
  convG = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{cor}{brownian motion ("BM") or autoregressive process ("AR") modeling the correlation
between the observations. NOTE that for this tool, the only accepted form is a string (ex: cor="BM(time)"),
which is different from the hlme call (ex: cor=BM(time)). NULL is still accepted.}

\item{idiag}{logical for the structure of the variance-covariance matrix of the random-effects.
If FALSE, a non structured matrix of variance-covariance is considered (by default).
If TRUE a diagonal matrix of variance-covariance is considered.}

\item{maxiter}{maximum number of iterations for the Marquardt iterative algorithm.}

\item{nproc}{the number cores for parallel computation. Default to 1 (sequential mode).}

\item{B_rand}{random effects "varcov" values to initiate the random effects model}

\item{convB}{optional threshold for the convergence criterion based on the
parameter stability. Used for the final MixedML model. By default, convB=0.0001.}

\item{convL}{optional threshold for the convergence criterion based on the
log-likelihood stability. Used for the final MixedML model. By default, convL=0.0001.}

\item{convG}{optional threshold for the convergence criterion based on the
derivatives. Used for the final MixedML model. By default, convG=0.0001.}
}
\value{
hlme_controls
}
\description{
Please see the \href{https://cecileproust-lima.github.io/lcmm/reference/hlme.html}{documentation}
of the \code{hlme} function of the \code{lcmm} package.
}
