% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{load_mixedml}
\alias{load_mixedml}
\title{Load a MixedML model}
\usage{
load_mixedml(mixedml_model_rds)
}
\arguments{
\item{mixedml_model_rds}{Name of the RDS fileNew data (same format as the one used for training)}
}
\value{
MixedMl model
}
\description{
Load a MixedML model
}
