% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{mixedml_ctrls}
\alias{mixedml_ctrls}
\title{Prepare the mixedml_controls}
\usage{
mixedml_ctrls(
  patience = 2,
  conv_thresh = 0.01,
  use_all_information_for_hlme_predictions = TRUE,
  no_random_value_as = NA,
  convB = 0.01,
  convL = 0.01,
  convG = 0.01
)
}
\arguments{
\item{patience}{Number of iterations without improvement before the training is stopped. Default: 2}

\item{conv_thresh}{Minimal difference of MSE to consider an improvement.
\code{conv_thresh=0.01} means an improvement of at least 1\% of the MSE is necessary. Default: 0.01}

\item{use_all_information_for_hlme_predictions}{If TRUE, then the hlme predictions
returned during the mixedML iterations will use all the information (past/future,
behaviour of the lcmm::predictY function).
If False, the predictions will only use the past information
(behaviour of the mixedml::.predict_random_hlme function).
Default: TRUE.}

\item{no_random_value_as}{value to use in the
"use_all_information_for_hlme_predictions==TRUE",
when the prediction is not possible .
This does not affect the prediction.
NA or 0. Default: NA}

\item{convB}{optional iterations models threshold for the convergence criterion based on the
parameter stability. Used during the MixedML iterations.
By default, convB=0.01.}

\item{convL}{optional threshold for the convergence criterion based on the
log-likelihood stability. Used during the MixedML iterations.
By default, convL=0.01.}

\item{convG}{optional threshold for the convergence criterion based on the
derivatives. Used during the MixedML iterations.
By default, convG=0.10.}
}
\value{
mixedml_controls
}
\description{
Prepare the mixedml_controls
}
