% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_conv}
\alias{plot_conv}
\title{Plot the (MSE) convergence of the MixedML training}
\usage{
plot_conv(model, ylog = TRUE)
}
\arguments{
\item{model}{Trained MixedML model}

\item{ylog}{Plot the y-value with a log scale. Defalut: TRUE.}
}
\value{
Convergence plot
}
\description{
Plot the (MSE) convergence of the MixedML training
}
