% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_conv_loglik}
\alias{plot_conv_loglik}
\title{Plot the log-likelihood of the random effect hlme during training}
\usage{
plot_conv_loglik(model)
}
\arguments{
\item{model}{Trained MixedML model}
}
\value{
Log-likelihood plot
}
\description{
Plot the log-likelihood of the random effect hlme during training
}
