% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_conv_mse}
\alias{plot_conv_mse}
\title{Plot the (MSE) convergence of the MixedML training and validation}
\usage{
plot_conv_mse(model, ylog = FALSE)
}
\arguments{
\item{model}{Trained MixedML model}

\item{ylog}{Plot the y-value with a log scale. Default: FALSE}
}
\value{
Convergence plot
}
\description{
Plot the (MSE) convergence of the MixedML training and validation
}
