% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot the MSE and loglikelihood convergence of the MixedML training and validation}
\usage{
plot_convergence(model, ylog_mse = FALSE)
}
\arguments{
\item{model}{Trained MixedML model}

\item{ylog_mse}{Plot the y-value with a log scale. Default: FALSE}
}
\value{
Convergence plot
}
\description{
Plot the MSE and loglikelihood convergence of the MixedML training and validation
}
