% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_last_iter}
\alias{plot_last_iter}
\title{Plot the prediction of a MixedML model}
\usage{
plot_last_iter(model, subject_nb_or_list, ylog = FALSE)
}
\arguments{
\item{model}{Trained MixedML model.}

\item{subject_nb_or_list}{Number of subjects to plot (randomly selected) or
list of subjects to plot.}

\item{ylog}{Plot the y-value with a log scale. Default: TRUE.}
}
\value{
Prediction plot of the model.
}
\description{
Plot the prediction of a MixedML model
}
