% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_loglik}
\alias{plot_loglik}
\title{Plot the log-likelihood of the random effect hlme during training}
\usage{
plot_loglik(model, ylog = TRUE)
}
\arguments{
\item{model}{Trained MixedML model}

\item{ylog}{Plot the y-value with a log scale. Default: TRUE.}
}
\value{
Log-likelihood plot
}
\description{
Plot the log-likelihood of the random effect hlme during training
}
