% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot individual predictions beside the corresponding true/target values}
\usage{
plot_predictions(model, data_pred, list_preds, ncols = 2)
}
\arguments{
\item{model}{Trained MixedML model.}

\item{data_pred}{Dataframe used to compute the predictions. It will be used to get the true/target value.}

\item{list_preds}{Named list of predictions to plot beside the true/target value. The names will be used in the
legend. Each prediction must be either a named vector or a single column dataframe, whose name/rowname must
correspond to the rownames of data_pred}

\item{ncols}{Numbre of columns to use for the grid plot. Default: 2}
}
\description{
Plot individual predictions beside the corresponding true/target values
}
