% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{plot_predictions_check}
\alias{plot_predictions_check}
\title{Plot the predictions of a MixedML model (with all or past informations) beside the corresponding true/target values.}
\usage{
plot_predictions_check(model, subject_list = NULL, ncols = 2)
}
\arguments{
\item{model}{Trained MixedML model.}

\item{subject_list}{List of subjects to plot (amongst the train/val dataset). If NULL, the predictions of all
individuals will be printed. Default: NULL}

\item{ncols}{Number of columns for the grid plot. Default: 2}
}
\value{
Predictions plot of the model.
}
\description{
Plot the predictions of a MixedML model (with all or past informations) beside the corresponding true/target values.
}
