% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{predict_ctrls}
\alias{predict_ctrls}
\title{Prepare the predict_controls}
\usage{
predict_ctrls(stateful = TRUE, reset = FALSE)
}
\arguments{
\item{stateful}{If True, Node state will be updated by this operation.}

\item{reset}{If True, Nodes states will be reset to zero before this operation.}
}
\value{
predict_controls
}
\description{
Please see the
\href{https://reservoirpy.readthedocs.io/en/latest/api/generated/reservoirpy.nodes.ESN.html#reservoirpy.nodes.ESN.run}{documentation}
of ReservoirPy
}
