% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.reservoir.R
\name{reservoir_mixedml}
\alias{reservoir_mixedml}
\title{Estimation of a mixedML model using Reservoir Computing}
\usage{
reservoir_mixedml(
  fixed_spec,
  random_spec,
  data,
  data_val = NULL,
  subject,
  time,
  mixedml_controls = mixedml_ctrls(),
  hlme_controls = hlme_ctrls(),
  esn_controls = esn_ctrls(),
  ensemble_controls = ensemble_ctrls(),
  fit_controls = fit_ctrls()
)
}
\arguments{
\item{fixed_spec}{two-sided linear formula object for the fixed-effects.
The response outcome is on the left of ~ and the covariates are separated by + on the right of ~.
(do not used extra formulation such as "x1*x3")}

\item{random_spec}{one-sided formula for the random-effects in the linear mixed model.
By default, an intercept is included. If no intercept, -1 should be the first term included.}

\item{data}{dataframe containing the variables named in \code{fixed_spec}, \code{random_spec}, \code{subject} and \code{time}.}

\item{data_val}{dataframe used for validation (control of over-training). If present, the earlystopping and
aborting procedures will use the MSE on this dataset. Default: NULL (no validation dataset).}

\item{subject}{name of the covariate representing the grouping structure, given as a string/character.}

\item{time}{name of the time variable, given as a string/character.}

\item{mixedml_controls}{controls specific to the MixedML model}

\item{hlme_controls}{controls specific to the HLME model}

\item{esn_controls}{controls specific to the ESN models}

\item{ensemble_controls}{controls specific to the Ensemble model}

\item{fit_controls}{controls specific to the ESN models fit
The default use the date at start in the format "mixedML-\%y\%m\%d-\%H\%M\%S"
(ex: mixedML-250709-100530)}
}
\value{
fitted MixedML model (best iteration)
}
\description{
The function fits a MixedML model which combines an Ensemble of Echo State Networks (Reservoir+Ridge Regression)
for the population part (fixed effects) and a linear mixed model for the individual part (random effects)
}
\examples{
set.seed(123)
data_mixedml <- data_mixedml[complete.cases(data_mixedml), ]
data_train <- data_mixedml[data_mixedml$ID < 9, ]
data_val <- data_mixedml[data_mixedml$ID >= 9, ]


model_reservoir <- reservoir_mixedml(
  fixed_spec = ym ~ 1 + x1 + x2 + x3,
  random_spec = ~ 1 + x1 + x2,
  data = data_train,
  data_val = data_val,
  subject = "ID",
  time = "time",
  # parameters for MixedML method
  mixedml_controls = mixedml_ctrls(
    earlystopping_controls = earlystopping_ctrls(min_mse_gain = 0.001, patience = 5),
    all_info_hlme_prediction = TRUE
  ),
  # controls (extra-parameters) for the hlme model
  hlme_controls = hlme_ctrls(maxiter = 50, idiag = TRUE),
  # controls (extra-parameters) for the ML model
  esn_controls = esn_ctrls(units = 20, ridge = 1e-4),
  ensemble_controls = ensemble_ctrls(seed_list = c(1, 2, 3, 4, 5)),
  fit_controls = fit_ctrls(warmup = 0)
)
}
