% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{save_mixedml}
\alias{save_mixedml}
\title{Save a MixedML model}
\usage{
save_mixedml(model, mixedml_model_rds, overwrite = FALSE)
}
\arguments{
\item{model}{Trained MixedML model}

\item{mixedml_model_rds}{Name of the RDS fileNew data (same format as the one used for training)}
}
\description{
Save a MixedML model
}
