% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_python_environment}
\alias{use_python_environment}
\title{Set the Python environment for mixedML}
\usage{
use_python_environment(environement_tool, environement_name)
}
\arguments{
\item{environement_tool}{Tool used to managge the environments. Can be one of "venv" or "conda"}

\item{environement_name}{Name of the environment.}
}
\value{
invisible NULL
}
\description{
Set the Python environment for mixedML
}
