% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R, R/data_roxygen.R
\docType{data}
\name{data_mixedml}
\alias{data_mixedml}
\title{Synthetic mixed effects dataset}
\format{
An object of class \code{data.frame} with 49 rows and 9 columns.
}
\usage{
data_mixedml

data_mixedml
}
\description{
This dataset is used to test and study the mixedML package
The fixed effects, random effects and corresponding mixed effects (with
and without noise) are given.
NAs are added in the independent variables and the response
}
\keyword{datasets}
