% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlme_tools.R
\name{.initiate_random_hlme}
\alias{.initiate_random_hlme}
\title{Initiate the HLME model}
\usage{
.initiate_random_hlme(
  target_name,
  random_spec,
  data,
  subject,
  var.time,
  hlme_controls
)
}
\arguments{
\item{random_spec}{random_spec}

\item{data}{data}

\item{subject}{subject}

\item{var.time}{var.time}

\item{hlme_controls}{hlme_controls}
}
\value{
HLME model
}
\description{
Initiate the HLME model
}
