% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlme_tools.R
\name{.initiate_random_hlme}
\alias{.initiate_random_hlme}
\title{Initiate the HLME model}
\usage{
.initiate_random_hlme(
  target_name,
  random_spec,
  data,
  subject,
  var.time,
  hlme_controls
)
}
\description{
Initiate the HLME model
}
