% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{ensemble_ctrls}
\alias{ensemble_ctrls}
\title{Prepare the ensemble_controls}
\usage{
ensemble_ctrls(seed_list = c(1, 2, 3), agg_func = "median", n_procs = 1)
}
\arguments{
\item{seed_list}{List of seeds used to generate the Reservoir. Default:  c(1, 2, 3)}

\item{agg_func}{Function used to aggregate the predictions of each ESN.
"mean" or "median". Default: "median"}

\item{n_procs}{Number of processor to use. 1 means no multiprocessing. Default: 1.}
}
\value{
ensemble_controls
}
\description{
Prepare the ensemble_controls
}
