% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_model.reservoir.R
\name{esn_ctrls}
\alias{esn_ctrls}
\title{Prepare the esn_controls}
\usage{
esn_ctrls(
  units = 100,
  lr = 1,
  sr = 0.1,
  ridge = 0,
  input_scaling = 1,
  feedback = FALSE,
  input_to_readout = FALSE
)
}
\arguments{
\item{units}{Number of reservoir units.}

\item{lr}{Neurons leak rate. Must be in \eqn{[0,1]}.}

\item{sr}{Spectral radius of recurrent weight matrix.}

\item{ridge}{Regularization parameter \eqn{\lambda}.}

\item{input_scaling}{Input gain. (So far only a float can be used).}

\item{feedback}{Is readout connected to reservoir through feedback?}

\item{input_to_readout}{If True, the input is directly fed to the readout.}
}
\value{
esn_controls
}
\description{
Please see the documentation of ReservoirPy for:
\itemize{
\item \href{https://reservoirpy.readthedocs.io/en/latest/api/generated/reservoirpy.nodes.Reservoir.html}{Reservoir}
\item \href{https://reservoirpy.readthedocs.io/en/latest/api/generated/reservoirpy.nodes.Ridge.html}{Ridge Regression}
}
}
