% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{esn_ctrls}
\alias{esn_ctrls}
\title{Prepare the esn_controls}
\usage{
esn_ctrls(units = 100, lr = 1, sr = 0.1, ridge = 0, feedback = FALSE)
}
\arguments{
\item{units}{Number of reservoir units.}

\item{lr}{Neurons leak rate. Must be in \link{0,1}.}

\item{sr}{Spectral radius of recurrent weight matrix.}

\item{ridge}{Regularization parameter \eqn{\lambda}.}

\item{feedback}{Is readout connected to reservoir through feedback?}
}
\value{
esn_controls
}
\description{
Please see the documentation of ReservoirPy for:
\itemize{
\item \href{https://reservoirpy.readthedocs.io/en/latest/api/generated/reservoirpy.nodes.Reservoir.html}{Reservoir}
\item \href{https://reservoirpy.readthedocs.io/en/latest/api/generated/reservoirpy.nodes.Ridge.html}{Ridge Regression}
}
}
