% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{fit_ctrls}
\alias{fit_ctrls}
\title{Prepare the fit_controls}
\usage{
fit_ctrls(warmup = 0, stateful = TRUE, reset = FALSE)
}
\arguments{
\item{warmup}{Number of timesteps to consider as warmup and discard at the beginning. Defalut: 0
of each timeseries before training.}

\item{stateful}{If True, Node state will be updated by this operation. Default: TRUE}

\item{reset}{If True, Nodes states will be reset to zero before this operation. Default: FALSE}
}
\value{
fit_controls
}
\description{
Please see the
\link[=1]{documentation}
of ReservoirPy
\link{1} https://reservoirpy.readthedocs.io/en/latest/api/generated/reservoirpy.nodes.ESN.html#reservoirpy.nodes.ESN.fit
}
