% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlme_tools.R
\name{hlme_ctrls}
\alias{hlme_ctrls}
\title{Prepare the hlme_controls}
\usage{
hlme_ctrls(idiag = FALSE, maxiter = 500, nproc = 1)
}
\value{
hlme_controls
}
\description{
Please see the parameters use in the \href{https://cecileproust-lima.github.io/lcmm/reference/hlme.html}{documentation}
of the \code{hlme} function of the \code{lcmm} package.
}
