% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{mixedml_ctrls}
\alias{mixedml_ctrls}
\title{Prepare the mixedml_controls}
\usage{
mixedml_ctrls(
  patience = 2,
  conv_thresh = 0.01,
  no_random_value_as = NA,
  convB = 0.01,
  convL = 0.01,
  convG = 0.01
)
}
\arguments{
\item{patience}{Number of iterations without improvement before the training is stopped. Default: 2}

\item{conv_thresh}{Minimal difference of MSE to consider an improvement.
\code{conv_thresh=0.01} means an improvement of at least 1\% of the MSE is necessary. Default: 0.01}

\item{no_random_value_as}{value to use during the training of the mixedML model
when a prediction of the random model is not possible (NA or 0).
This does not affect the prediction.
Default: NA}

\item{convB}{optional iterations models threshold for the convergence criterion based on the
parameter stability. Used during the MixedML iterations.
By default, convB=0.01.}

\item{convL}{optional threshold for the convergence criterion based on the
log-likelihood stability. Used during the MixedML iterations.
By default, convL=0.01.}

\item{convG}{optional threshold for the convergence criterion based on the
derivatives. Used during the MixedML iterations.
By default, convG=0.10.}
}
\value{
mixedml_controls
}
\description{
Prepare the mixedml_controls
}
