% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{mixedml_ctrls}
\alias{mixedml_ctrls}
\title{Prepare the mixedml_controls}
\usage{
mixedml_ctrls(
  earlystopping_controls = earlystopping_ctrls(),
  aborting_controls = aborting_ctrls(),
  all_info_hlme_prediction = TRUE,
  convB = 0.01,
  convL = 0.01,
  convG = 0.01
)
}
\arguments{
\item{earlystopping_controls}{controls specific to the early stopping criterion. Please see earlystopping_ctrls().}

\item{aborting_controls}{controls specific to the aborting criterion. Please see aborting_ctrls().}

\item{all_info_hlme_prediction}{boolean to choose if all the information
(past, present, future) is used for the hlme prediction (TRUE) or if only the past
information is used (FALSE). Default: TRUE}

\item{convB}{optional iterations models threshold for the convergence criterion based on the
parameter stability. Used during the MixedML iterations to speed up the HLME model training.
By default, convB=0.01.}

\item{convL}{optional threshold for the convergence criterion based on the
log-likelihood stability. Used during the MixedML iterations to speed up the HLME model training.
By default, convL=0.01.}

\item{convG}{optional threshold for the convergence criterion based on the
derivatives. Used during the MixedML iterations to speed up the HLME model training.
By default, convG=0.01.}
}
\value{
mixedml_controls
}
\description{
Prepare the mixedml_controls
}
