% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{mixedml_ctrls}
\alias{mixedml_ctrls}
\title{Prepare the mixedml_controls}
\usage{
mixedml_ctrls(patience = 2, conv_ratio_thresh = 0.01)
}
\arguments{
\item{patience}{Number of iterations without improvement before the training is stopped. Default: 2}

\item{conv_ratio_thresh}{Ratio of improvement of the MSE to consider an improvement.
\code{conv_ratio_thresh=0.01} means an improvement of at least 1\% of the MSE is necessary. Default: 0.01}
}
\value{
mixedml_controls
}
\description{
Prepare the mixedml_controls
}
