% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{predict}
\alias{predict}
\title{Predict using a fitted model and new data}
\usage{
predict(model, data)
}
\arguments{
\item{model}{Trained MixedML model}

\item{data}{New data (same format as the one used for training)}
}
\value{
prediction
}
\description{
Predict using a fitted model and new data
}
