% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedml.R
\name{reservoir_mixedml}
\alias{reservoir_mixedml}
\title{MixedML model with Reservoir Computing}
\usage{
reservoir_mixedml(
  fixed_spec,
  random_spec,
  data,
  subject,
  time,
  mixedml_controls = mixedml_ctrls(),
  hlme_controls = hlme_ctrls(),
  esn_controls = esn_controls(),
  ensemble_controls = ensemble_controls(),
  fit_controls = fit_controls(),
  predict_controls = predict_controls()
)
}
\arguments{
\item{fixed_spec}{two-sided linear formula object for the fixed-effects.
The response outcome is on the left of ~ and the covariates are separated by + on the right of ~.}

\item{random_spec}{two-sided formula for the random-effects in the linear mixed model.
The response outcome is on the left of ~ and the covariates are separated by + on the right of ~.
By default, an intercept is included. If no intercept, -1 should be the first term included.}

\item{data}{dataframe containing the variables named in \code{fixed_spec}, \code{random_spec}, \code{subject} and \code{time}.}

\item{subject}{name of the covariate representing the grouping structure, given as a string/character.}

\item{time}{name of the time variable, given as a string/character.}

\item{mixedml_controls}{controls specific to the MixedML model}

\item{hlme_controls}{controls specific to the HLME model}

\item{esn_controls}{controls specific to the ESN models}

\item{ensemble_controls}{controls specific to the Ensemble model}

\item{fit_controls}{controls specific to the ESN models fit}

\item{predict_controls}{controls specific to the ESN models prediction}
}
\value{
fitted MixedML model
}
\description{
Generate and fit a MixedML model using an Ensemble of Echo State Networks (Reservoir+Ridge Regression)
to fit the fixed effects.
}
